Move cell location database to subpackage
